+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 * fix processing list with nested unquote-splicing [#362](https://github.com/LIPS-scheme/lips/issues/362)
 * fix exception during error from eval [#362](https://github.com/LIPS-scheme/lips/issues/362)
 * allow to catch exceptions from eval
+* fix else keyword as only expression in cond [#366](https://github.com/LIPS-scheme/lips/issues/366)
 
 ## 1.0.0-beta.19
 ### Breaking